# wwwroot/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "DOLFIN API",
    "version": "1.0.0",
    "description": "DOLFIN Financial Intelligence and Portfolio Management API Documentation"
  },
  "servers": [
    {
      "url": "https://ensaraslannn-001-site1.ctempurl.com",
      "description": "Production Server"
    },
    {
      "url": "https://localhost:7109",
      "description": "Local Development Server"
    }
  ],
  "paths": {
    "/api/account/login": {
      "post": {
        "tags": ["Account"],
        "summary": "User Login",
        "description": "Authenticates user credentials and generates a valid JWT Bearer Token.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["username", "password"],
                "properties": {
                  "username": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  }
                }
              },
              "example": {
                "username": "exampleusername",
                "password": "examplepassword"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful. Returns NewUserDto with JWT Token."
          },
          "401": {
            "description": "Unauthorized. Invalid username or password!"
          }
        }
      }
    },
    "/api/account/register": {
      "post": {
        "tags": ["Account"],
        "summary": "Register New User",
        "description": "Creates a new AppUser instance in the database and assigns the default 'User' role.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["username", "email", "password"],
                "properties": {
                  "username": {
                    "type": "string",
                    "example": "example123"
                  },
                  "email": {
                    "type": "string",
                    "format": "email",
                    "example": "example@example.com"
                  },
                  "password": {
                    "type": "string",
                    "format": "password",
                    "example": "Example123!"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Registration successful. Returns user details with an active token."
          },
          "500": {
            "description": "Internal Server Error. Identity rule validation failed."
          }
        }
      }
    },
    "/api/comment": {
      "get": {
        "tags": ["Comment"],
        "summary": "List All Comments",
        "description": "Retrieves comments filtered by QueryObject parameters. Requires authentication.",
        "security": [
          {
            "Bearer": []
          }
        ],
        "parameters": [
          {
            "name": "Symbol",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Filter comments by stock symbol",
            "example": "EXMPL"
          },
          {
            "name": "IsDescending",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean"
            },
            "description": "Sort order for comment creation dates"
          }
        ],
        "responses": {
          "200": {
            "description": "Success. Returns a list of CommentDto objects."
          },
          "400": {
            "description": "Bad Request. Invalid ModelState."
          }
        }
      }
    },
    "/api/comment/{id}": {
      "get": {
        "tags": ["Comment"],
        "summary": "Get Comment By ID",
        "description": "Retrieves a single comment based on its unique SQL database ID. Open to public.",
        "security": [
          {
            "Bearer": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "The unique identifier of the comment"
          }
        ],
        "responses": {
          "200": {
            "description": "Success. Returns the requested comment object."
          },
          "404": {
            "description": "Not Found. Comment does not exist."
          }
        }
      },
      "put": {
        "tags": ["Comment"],
        "summary": "Update Existing Comment",
        "description": "Updates the content of the specified comment using the request body data.",
        "security": [
          {
            "Bearer": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "The ID of the comment to update"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["title", "content"],
                "properties": {
                  "title": {
                    "type": "string"
                  },
                  "content": {
                    "type": "string"
                  }
                }
              },
              "example": {
                "title": "exampletitle",
                "content": "examplecontent"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Update successful. Returns the modified CommentDto."
          },
          "404": {
            "description": "Not Found. Comment not found in database."
          }
        }
      },
      "delete": {
        "tags": ["Comment"],
        "summary": "Delete Comment",
        "description": "Removes the comment record from SQL Server and returns the deleted entity for confirmation.",
        "security": [
          {
            "Bearer": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "The ID of the comment to delete"
          }
        ],
        "responses": {
          "200": {
            "description": "Delete successful. Returns the deleted raw entity model."
          },
          "404": {
            "description": "Not Found. Comment does not exist."
          }
        }
      }
    },
    "/api/comment/{stockId}": {
      "post": {
        "tags": ["Comment"],
        "summary": "Add Comment to Stock",
        "description": "Creates a new comment linked to a specific stock entity for the currently authenticated user.",
        "security": [
          {
            "Bearer": []
          }
        ],
        "parameters": [
          {
            "name": "stockId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "The unique SQL ID of the target stock",
            "example": 1
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["title", "content"],
                "properties": {
                  "title": {
                    "type": "string"
                  },
                  "content": {
                    "type": "string"
                  }
                }
              },
              "example": {
                "title": "exampletitle",
                "content": "examplecontent"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created. Comment successfully added, location header updated via CreatedAtAction."
          },
          "400": {
            "description": "Bad Request. Stock does not exist or invalid input values."
          }
        }
      }
    },
    "/api/portfolio": {
      "get": {
        "tags": ["Portfolio"],
        "summary": "Get User Portfolio",
        "description": "Resolves current user context via claims and lists all associated stock investments.",
        "security": [
          {
            "Bearer": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success. Returns the authenticated user's portfolio list."
          }
        }
      },
      "post": {
        "tags": ["Portfolio"],
        "summary": "Add Stock to Portfolio",
        "description": "Buys a stock, deducts funds from the user's wallet, updates average cost, and logs the transaction.",
        "security": [
          {
            "Bearer": []
          }
        ],
        "parameters": [
          {
            "name": "symbol",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The equity symbol to purchase (e.g., TSLA)",
            "example": "MSFT"
          },
          {
            "name": "quantity",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "The number of shares to buy (must be greater than 0)",
            "example": 5
          }
        ],
        "responses": {
          "200": {
            "description": "Success. Stock purchased successfully. Returns remaining wallet balance."
          },
          "400": {
            "description": "Bad Request. Insufficient funds, invalid quantity, or stock not found."
          },
          "401": {
            "description": "Unauthorized. User context or valid token not found."
          }
        }
      }
    },
    "/api/portfolio/sell": {
      "post": {
        "tags": ["Portfolio"],
        "summary": "Sell Stock from Portfolio",
        "description": "Sells a specified quantity of a stock, adds revenue to the user's wallet, and logs the transaction.",
        "security": [
          {
            "Bearer": []
          }
        ],
        "parameters": [
          {
            "name": "symbol",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The equity symbol to sell",
            "example": "AAPL"
          },
          {
            "name": "quantity",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "The number of shares to sell",
            "example": 2
          }
        ],
        "responses": {
          "200": {
            "description": "Success. Stock sold successfully. Returns updated wallet balance."
          },
          "400": {
            "description": "Bad Request. Insufficient stock quantity or invalid input."
          }
        }
      }
    },
    "/api/portfolio/deposit": {
      "post": {
        "tags": ["Portfolio"],
        "summary": "Deposit Funds to Wallet",
        "description": "Adds the specified amount to the authenticated user's wallet balance and logs a DEPOSIT transaction.",
        "security": [
          {
            "Bearer": []
          }
        ],
        "parameters": [
          {
            "name": "amount",
            "in": "query",
            "required": true,
            "schema": {
              "type": "number",
              "format": "decimal"
            },
            "description": "The amount of cash to deposit into the wallet",
            "example": 1000
          }
        ],
        "responses": {
          "200": {
            "description": "Success. Funds deposited successfully. Returns updated wallet balance."
          },
          "400": {
            "description": "Bad Request. Deposit amount must be greater than 0."
          }
        }
      }
    },
    "/api/portfolio/withdraw": {
      "post": {
        "tags": ["Portfolio"],
        "summary": "Withdraw Funds from Wallet",
        "description": "Deducts the specified amount from the authenticated user's wallet balance and logs a WITHDRAW transaction.",
        "security": [
          {
            "Bearer": []
          }
        ],
        "parameters": [
          {
            "name": "amount",
            "in": "query",
            "required": true,
            "schema": {
              "type": "number",
              "format": "decimal"
            },
            "description": "The amount of cash to withdraw",
            "example": 500
          }
        ],
        "responses": {
          "200": {
            "description": "Success. Funds withdrawn successfully. Returns updated wallet balance."
          },
          "400": {
            "description": "Bad Request. Insufficient funds or invalid amount."
          }
        }
      }
    },
    "/api/stock": {
      "get": {
        "tags": ["Stock"],
        "summary": "List All Stocks",
        "description": "Fetches equities matching filter criteria from query parameters. Open endpoint.",
        "security": [
          {
            "Bearer": []
          }
        ],
        "parameters": [
          {
            "name": "Symbol",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Filter by ticker symbol (e.g., MSFT)"
          },
          {
            "name": "CompanyName",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Filter by legal corporate title"
          },
          {
            "name": "PageNumber",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 1
            },
            "description": "The page number to retrieve"
          },
          {
            "name": "PageSize",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 20
            },
            "description": "The number of items per page"
          },
          {
            "name": "SortBy",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Field to sort by (e.g., Symbol, MarketCap)"
          },
          {
            "name": "IsDescending",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false
            },
            "description": "Set to true for descending order"
          }
        ],
        "responses": {
          "200": {
            "description": "Success. Returns a parsed list of StockDto entries."
          }
        }
      },
      "post": {
        "tags": ["Stock"],
        "summary": "Create New Stock Asset",
        "description": "Maps incoming JSON data into a clean ID-less stock entity and persists it into SQL server tables.",
        "security": [
          {
            "Bearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "symbol",
                  "companyName",
                  "purchase",
                  "lastDiv",
                  "industry",
                  "marketCap"
                ],
                "properties": {
                  "symbol": {
                    "type": "string"
                  },
                  "companyName": {
                    "type": "string"
                  },
                  "purchase": {
                    "type": "number",
                    "format": "decimal"
                  },
                  "lastDiv": {
                    "type": "number",
                    "format": "decimal"
                  },
                  "industry": {
                    "type": "string"
                  },
                  "marketCap": {
                    "type": "integer",
                    "format": "int64"
                  }
                }
              },
              "example": {
                "symbol": "EXMPL",
                "companyName": "Example",
                "purchase": 0,
                "lastDiv": 0,
                "industry": "Example",
                "marketCap": 0
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created. Asset initialized with a REST location pointer framework."
          }
        }
      }
    },
    "/api/stock/{id}": {
      "get": {
        "tags": ["Stock"],
        "summary": "Get Stock Details By ID",
        "description": "Finds a specific corporate asset entity by its unique identifier primary key.",
        "security": [
          {
            "Bearer": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "The target stock database ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Success. Corporate stock profile loaded."
          },
          "404": {
            "description": "Not Found. Asset code does not exist in our index system."
          }
        }
      },
      "put": {
        "tags": ["Stock"],
        "summary": "Update Stock Profile",
        "description": "Updates financial ratios, caps, and names on the targeted table row asset.",
        "security": [
          {
            "Bearer": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "symbol",
                  "companyName",
                  "purchase",
                  "lastDiv",
                  "industry",
                  "marketCap"
                ],
                "properties": {
                  "symbol": {
                    "type": "string"
                  },
                  "companyName": {
                    "type": "string"
                  },
                  "purchase": {
                    "type": "number",
                    "format": "decimal"
                  },
                  "lastDiv": {
                    "type": "number",
                    "format": "decimal"
                  },
                  "industry": {
                    "type": "string"
                  },
                  "marketCap": {
                    "type": "integer",
                    "format": "int64"
                  }
                }
              },
              "example": {
                "symbol": "EXMPL",
                "companyName": "Example",
                "purchase": 0,
                "lastDiv": 0,
                "industry": "Example",
                "marketCap": 0
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Update complete. Modified records applied successfully."
          },
          "404": {
            "description": "Not Found. Stock target missing from records."
          }
        }
      },
      "delete": {
        "tags": ["Stock"],
        "summary": "Delete Stock From System",
        "description": "Deletes the stock row from SQL server and throws back a 204 NoContent server status code.",
        "security": [
          {
            "Bearer": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content. Row successfully purged, response contains empty body."
          },
          "404": {
            "description": "Not Found. Delete operation aborted because ID is missing."
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Please input a valid JWT Token to access DOLFIN endpoints."
      }
    }
  }
}